refactor: PlaceLikeRepository 패키지 변경
