+++ backend/spring-routie/src/main/java/routie/business/placelike/domain/PlaceLikeRepository.java
@@ -1,8 +1,7 @@
-package routie.business.placelike;
+package routie.business.placelike.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import routie.business.placelike.domain.PlaceLike;
 
 @Repository
 public interface PlaceLikeRepository extends JpaRepository<PlaceLike, Long> {
